[week3]게임 모드 선택
