+++ week3/card_flip/src/App.jsx
@@ -10,7 +10,7 @@ function App() {
   // 상수파일 -> 데이터 랜더링
 
   //심화 리셋 버튼 클릭 시 랜덤 => 리셋에 shuffleCards()함수 온클릭으로 넣기
-  //컴포 분리 => 헤더->한 번더 분리? (제목/리셋 버튼/ 스코어) =>  cardBody (난이도 선택/ card리스트)
+  //컴포 분리 => 헤더->한 번더 분리? (제목/리셋 버튼/ 스코어) =>  EasyMode (난이도 선택/ card리스트)
   return (
     <>
       <Header />
