+++ week3/card_flip/src/component/Header.jsx
@@ -4,8 +4,6 @@ function Header() {
   return (
     <div>
       <StTitle>💸이거 맞추면 나 사주는 거다 ?💸</StTitle>
-
-      <button>reset</button>
     </div>
   );
 }
